requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.11
Deprecated==1.2.13
Django==3.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-rosetta==0.9.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
packaging==21.3
polib==1.1.1
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
wrapt==1.13.3
